Distillers now return a list of documents to index
Some distillers really want to return multiple documents for a given
URL. This is now the standard case, and all distillers return a list of
documents for indexing, even if it's just a singleton list.

As part of the change, a document is no longer a string, but a dict with
a `url` and `blob`. This is particularly helpful because the canonical
URL for a document may not be related to the URL you actually supplied.
A good example would be the ticket API - a user cares about getting a
URL to RT, not something at ticket.api.anchor.com.au