Switch to using a Sorted Set instead of plain Set
Sets are almost exactly what we need, except that we have no way of
easily tracking timestamps of inserted elements. This isn't a huge deal
in and of itself, but we need to be able to monitor the ages of elements
in the indexing queue, because queue_len is a shitty metric for health
of the service.

Sorted sets let us specify arbitrary scores, which in this case are unix
timestamps. We now get order-by-age for free, and we can pluck out the
lowest ranked element to find the oldest URL in the box. Unfortunately
we lose the best properties of Sets, like O(1) performance and the
ability to atomically pop elements. :(