Whoops, should really use StrictRedis everywhere.
StrictRedis exists because of a regression from earlier versions of the
library. Specifically with respect to ZADD, redis-py used to take the
(value,score), which is the opposite of official Redis order, which is
(score,value).

StrictRedis implements the correct ordering, but with a different name
so as not to break existing code. We will use the correct ordering.