Fix programming errors in the derivation of RT digests
So yeah, first_post was being taken as a reference instead of a copy of
messages[0]. When I trimmed the content of first_post, it was spoiling
the original copy of the message.

On a somewhat related note, I baked all_message_lines into a list from
the generator, so I could reuse it for debugging and whatnot, might as
well leave it that way.