Rework the provsys auditlog watching completely
Now that we can ask for ID ranges from provsys, make use of that. Keep
track of the auditlog entries that we *have* seen, and record timestamps
so we can tell how far we are behind reality.

Instead of polling for entries that may not be there, we just look for
anything newer that what we know we've seen.