Fix up ES highlighting parameters in the query
The excerpt field was being given too much priority, always being
included even if it didn't contain any search terms. Provsys docs
always contain the search term, but things like wiki pages and RT
tickets often won't.

As an aside, documents arguably shouldn't need an excerpt field now that
ES will extract and highlight matching fragments for us.

These changes mean that the excerpt-highlight will only be populated if
there's a real match. The blob field remains as a fallback and *will*
always contain content, although I *think* it's possible for it to not
be highlighted (if the no_match_size kicks in).

The renderer prefers to display the excerpt-highlight if it exists,
otherwise it displays whatever the blob-highlight has to offer.