Resolves issue #6, no longer return results with unhighlighted snippets
ES was observed to be returning unescaped snippets even though I'd
directed the highlighter to run output through the HTML encoder before
slapping the highlight tags on. I eventually tracked this down to the
fact that the encoder only kicks in when highlighting actually takes
place.

no_match_size was being specified for the blob field, forcing ES to
return output even in the absence of a highlightable match, and the
encoder is skipped, meaning that the client sees unescaped HTML.

Not using no_match_size fixes this, which I've done. The client will
make use of the blob field itself, so it's not imperative that ES return
highlight data.