Setup base Distiller class to accept indexer_url on instantiation (WIP)
The aim is to make indexer_url available to each distiller, which can be
used in case it needs to trigger indirectly-related indexing/deletion
downstream without getting in the way of the current indexing operation.

This is cleaner than pulling in the ElasticSearch library and allowing
distillers to call ES directly.

The distiller mechanism should probably be reworked to be properly OO,
making each distiller a subclass of the base class.